// packages/core/src/preset/preset.d.ts
export interface Preset {
    /**
     * Internal unique identifier for the preset
     */
    id: string;
    name: string;
    description: string;
    author: string;
    createdAt: Date;
    updatedAt: Date;
    version: string;
    systemPrompt: string;
    enabledMcps?: EnabledMcp[];
    llmBridgeName: string;
    llmBridgeConfig: Record<string, any>;
}
export interface EnabledMcp {
    name: string;
    version?: string;
    enabledTools: string[];
    enabledResources: string[];
    enabledPrompts: string[];
}
//# sourceMappingURL=preset.d.ts.map